requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-docx>=1.1.0
pyinstaller>=6.0.0 